implementação php

+++ progs/docker-remove.sh
@@ -5,18 +5,10 @@ if [ -v $2 ]; then
   exit;
 fi
 
-if docker ps | grep "$2" &> /dev/null; then
+if docker ps -a | grep "$2" &> /dev/null; then
   docker stop "$2" && docker rm "$2";
-  echo ''
-  bash ${path_dir}/progs/success.sh 'Container excluído com sucesso!'
-else
-  bash ${path_dir}/progs/error.sh  "Container não encontrado!!";
-  exit;
-fi
-
-if docker ps | grep "$2" &> /dev/null; then
   docker image rm "$2" --force;
   echo ''
-  bash ${path_dir}/progs/success.sh 'Imagem excluído com sucesso!'
+  bash ${path_dir}/progs/success.sh 'Container excluído com sucesso!'
 fi
